[refactor] change file directory
