+++ src/main/java/umc/hackathon/domain/entity/Plant.java
@@ -1,4 +1,4 @@
-package umc.hackathon.domain;
+package umc.hackathon.domain.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
